@@ -1,7 +1,9 @@
 {
   "name": "client",
   "version": "0.1.0",
-  "private": true,
+  "engines": {
+    "node": "16.13.0"
+  },
   "dependencies": {
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
@@ -11,16 +13,16 @@
     "classnames": "^2.2.6",
     "materialize-css": "^1.0.0-rc.2",
     "moment": "^2.29.0",
-    "node-sass": "^4.14.1",
-    "react": "^16.13.1",
+    "path-to-regexp": "^1.7.0",
+    "react": "^16.14.0",
     "react-chartjs-2": "^2.10.0",
-    "react-dom": "^16.13.1",
-    "react-redux": "^7.2.1",
+    "react-dom": "^16.14.0",
+    "react-redux": "^7.2.6",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.3",
-    "redux": "^4.0.5",
-    "path-to-regexp": "^1.7.0"
+    "react-table": "^7.7.0",
+    "redux": "^4.0.5"
   },
   "scripts": {
     "start": "react-scripts start",
