@@ -3,6 +3,9 @@
   "version": "1.0.0",
   "description": "Transfers app for FM",
   "main": "server.js",
+  "engines": {
+    "node": "16.13.0"
+  },
   "scripts": {
     "build": "cd client && npm run build",
     "install-client": "cd client && npm install",
@@ -10,7 +13,7 @@
     "start": "node server.js",
     "server": "nodemon server.js",
     "client": "npm start --prefix client",
-    "dev": "concurrently \"npm run server\" \"npm run client\""
+    "dev": "concurrently \"npm:start\" \"npm:client\""
   },
   "keywords": [
     "mongoose",
@@ -28,12 +31,12 @@
     "express": "^4.17.1",
     "js-cookie": "^2.2.1",
     "jsonwebtoken": "^8.5.1",
+    "materialize-css": "^1.0.0-rc.2",
     "mongoose": "^5.10.6",
-    "nouislider": "^14.6.2",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-nouislider": "^2.0.1",
+    "react": "^16.14.0",
+    "react-dom": "^16.14.0",
     "react-redux": "^7.2.1",
+    "react-window": "^1.8.6",
     "redux": "^4.0.5",
     "redux-thunk": "^2.3.0",
     "wnumb": "^1.2.0"
